requirements.txt
absl-py==0.11.0
alembic==1.4.3
APScheduler==3.6.3
asn1crypto==0.24.0
astor==0.8.1
astroid==2.1.0
asttokens==1.1.13
automationhat==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
cached-property==1.5.2
cachetools==4.1.1
Cap1xxx==0.1.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==3.3.1
decorator==4.4.2
dnspython==2.0.0
docutils==0.14
dominate==2.6.0
drumhat==0.1.0
email-validator==1.1.2
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fourletterphat==0.1.0
gast==0.2.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gpiozero==1.5.1
grpcio==1.33.2
h5py==2.9.0
html5lib==1.0.1
idna==2.10
importlib-metadata==3.1.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.3.1
logilab-common==1.4.2
lxml==4.3.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
microdotphat==0.2.1
mock==4.0.2
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.1
mysql-client==0.0.1
mysqlclient==2.0.1
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opencv-python==3.4.6.27
opt-einsum==3.3.0
pantilthat==0.0.7
paramiko==2.7.2
parso==0.3.1
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.44
Pillow==5.4.1
protobuf==3.14.0
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
PyMySQL==0.10.1
PyNaCl==1.4.0
pyngrok==5.0.1
pyOpenSSL==19.0.0
pyserial==3.4
python-apt==1.8.4.1
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==13.1
pytz==2020.5
pyxdg==0.25
PyYAML==5.4.1
rainbowhat==0.1.0
requests==2.2.1
requests-oauthlib==1.3.0
responses==0.9.0
roman==2.0.0
RPi.GPIO==0.7.0
rsa==4.6
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.2.0
simplejson==3.16.0
six==1.15.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.4
SQLAlchemy==1.3.22
ssh-import-id==5.7
sshtunnel==0.4.0
tensorboard==2.0.2
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
thonny==3.3.0
tornado==6.1
touchphat==0.0.1
twython==3.7.0
typed-ast==1.3.1
tzlocal==2.1
unicornhathd==0.0.4
urllib3==1.26.2
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.0
